Move visit_nodes and visit_node_generations to runtime package
