+++ cubed/extensions/tqdm.py
@@ -3,8 +3,8 @@ import sys
 
 from toolz import map
 
-from cubed.core.array import Callback
-from cubed.core.plan import visit_nodes
+from cubed.runtime.pipeline import visit_nodes
+from cubed.runtime.types import Callback
 
 
 class TqdmProgressBar(Callback):
