+++ cubed/runtime/executors/coiled.py
@@ -4,7 +4,7 @@ import coiled
 from networkx import MultiDiGraph
 
 from cubed.core.array import Spec
-from cubed.core.plan import visit_nodes
+from cubed.runtime.pipeline import visit_nodes
 from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import execution_stats, handle_callbacks
