+++ cubed/runtime/executors/modal.py
@@ -9,7 +9,7 @@ from networkx import MultiDiGraph
 from tenacity import retry, retry_if_exception_type, stop_after_attempt
 
 from cubed.core.array import Spec
-from cubed.core.plan import visit_nodes
+from cubed.runtime.pipeline import visit_nodes
 from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import execute_with_stats, handle_callbacks
