+++ cubed/runtime/executors/modal_async.py
@@ -10,7 +10,6 @@ from networkx import MultiDiGraph
 from tenacity import retry, retry_if_exception_type, stop_after_attempt
 
 from cubed.core.array import Spec
-from cubed.core.plan import visit_node_generations, visit_nodes
 from cubed.runtime.executors.asyncio import async_map_unordered
 from cubed.runtime.executors.modal import (
     Container,
@@ -18,6 +17,7 @@ from cubed.runtime.executors.modal import (
     run_remotely,
     stub,
 )
+from cubed.runtime.pipeline import visit_node_generations, visit_nodes
 from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import handle_callbacks
